+++ src/compas_rhino/geometry/brep/brep.py
@@ -65,6 +65,9 @@ class RhinoBrep(Brep):
         super(RhinoBrep, self).__init__()
         self._brep = brep or Rhino.Geometry.Brep()
 
+    def __deepcopy__(self, *args, **kwargs):
+        return self.copy()
+
     # ==============================================================================
     # Data
     # ==============================================================================
@@ -87,6 +90,17 @@ class RhinoBrep(Brep):
         for f_data in data["faces"]:
             RhinoBrepFace.from_data(f_data, builder)
         self._brep = builder.result
+
+    def copy(self, cls=None):
+        """Creates a deep-copy of this Brep using the native Rhino.Geometry.Brep copying mechanism.
+
+        Returns
+        -------
+        :class:`~compas_rhino.geometry.RhinoBrep`
+
+        """
+        # Avoid reconstruction when just copying. for sake of efficiency and stability
+        return RhinoBrep.from_native(self._brep.DuplicateBrep())
 
     # ==============================================================================
     # Properties
@@ -219,7 +233,7 @@ class RhinoBrep(Brep):
         Parameters
         ----------
         trimming_plane : :class:`~compas.geometry.Frame` or :class:`~compas.geometry.Plane`
-            The frame or plane to use when trimming.
+            The frame or plane to use when trimming. The discarded bit is in the direction of the frame's normal.
 
         tolerance : float
             The precision to use for the trimming operation.
@@ -232,7 +246,6 @@ class RhinoBrep(Brep):
         if isinstance(trimming_plane, Plane):
             trimming_plane = Frame.from_plane(trimming_plane)
         rhino_frame = frame_to_rhino(trimming_plane)
-        rhino_frame.Flip()
         results = self._brep.Trim(rhino_frame, tolerance)
         if not results:
             raise BrepTrimmingError("Trim operation ended with no result")