+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 * Updated workflows to v2.
+* Changed deepcopy of `RhinoBrep` to use the native `Rhino.Geometry` mechanism.
+* The normal of the cutting plane is no longer flipped in `compas_rhino.geometry.RhinoBrep`.
 
 ### Removed